+++ src/jax_fdm/constraints/edge/force.py
@@ -10,9 +10,8 @@ class EdgeForceConstraint(EdgeConstraint):
     """
     Constraints the force of an edge between a lower and an upper bound.
     """
-    @staticmethod
-    def constraint(eqstate: EquilibriumState, index: Int[Array, ""]) -> Float[Array, "1"]:
+    def constraint(self, eqstate: EquilibriumState, index: Int[Array, ""]) -> Float[Array, ""]:
         """
-        Returns the length of an edge from an equilibrium state.
+        Returns the force of an edge from an equilibrium state.
         """
-        return eqstate.forces[index, :]
+        return eqstate.forces[index, 0]