+++ src/jax_fdm/constraints/network/network.py
@@ -10,7 +10,7 @@ class NetworkConstraint(Constraint):
     """
     Base class for all constraints that pertain to all the edges or all the nodes of a network.
     """
-    def __init__(self, bound_low: float | Float[Array, "..."] | None, bound_up: float | Float[Array, "..."] | None):
+    def __init__(self, bound_low: float | Float[Array, "..."] | None, bound_up: float | Float[Array, "..."] | None) -> None:
         super().__init__(key=-1, bound_low=bound_low, bound_up=bound_up)
 
     def index_from_model(self, model: EquilibriumModel, structure: EquilibriumStructure) -> int: